update: new design admin view dan admin layout by category
