+++ src/services/useStatisticStore.ts
@@ -359,7 +359,7 @@ export const useStatisticStore = defineStore({
             })
       }
    },
-   
+
    getters: {
       getTotalAbsensi(state: StatisticStoreState) {
          return state
