--- src/views/admin/AdminDashboard.vue
@@ -1,16 +1,0 @@
-<template>
-  <div>Admin Dashboard</div>
-</template>
-
-<script lang="ts">
-import { defineComponent } from "vue";
-
-export default defineComponent({
-  setup() {
-    return {};
-  },
-});
-</script>
-
-<style scoped>
-</style>
